okay, vcr making multiple requests now
