+++ gridstatus/tests/test_isone_api.py
@@ -1,62 +1,23 @@
-import json
-import os
-from urllib.parse import parse_qs, urlparse
-
 import numpy as np
 import pandas as pd
 import pytest
-import vcr
 
 from gridstatus.isone_api.isone_api import ISONEAPI, ZONE_LOCATIONID_MAP
+from tests.vcr_utils import record_mode, setup_vcr
 
+# NOTE(Kladar): Since these support date ranges and make multiple API calls in a single get
+# request, vcr needs to combine them into a single cassette.
+DATE_RANGE_ENDPOINTS = [
+    "realtimehourlydemand",
+    "dayaheadhourlydemand",
+    "hourlyloadforecast",
+    "reliabilityregionloadforecast",
+]
 
-def combine_requests(requests: list[vcr.request.Request]) -> str:
-    combined = []
-    for request in requests:
-        combined.append(
-            {
-                "method": request.method,
-                "uri": request.uri,
-                "body": request.body,
-                "headers": dict(request.headers),
-            },
-        )
-    return json.dumps(combined)
-
-
-def before_record_callback(request: vcr.request.Request) -> vcr.request.Request:
-    parsed_url = urlparse(request.uri)
-    path_parts = parsed_url.path.split("/")
-    date_range_endpoints = [
-        "realtimehourlydemand",
-        "dayaheadhourlydemand",
-        "hourlyloadforecast",
-        "reliabilityregionloadforecast",
-    ]
-
-    if any(endpoint in path_parts for endpoint in date_range_endpoints):
-        query_params = parse_qs(parsed_url.query)
-        if "date" in query_params and "end" in query_params:
-            key = f"{query_params['date'][0]}_{query_params['end'][0]}"
-            if not hasattr(before_record_callback, "requests"):
-                before_record_callback.requests = {}
-            if key not in before_record_callback.requests:
-                before_record_callback.requests[key] = []
-            before_record_callback.requests[key].append(request)
-            combined_body = combine_requests(before_record_callback.requests[key])
-            request.body = combined_body.encode("utf-8")
-
-    return request
-
-
-# NOTE(Kladar): Set VCR_RECORD_MODE to "all" to update the fixtures as an integration test,
-# say on a weekly or monthly job.
-record_mode = os.environ.get("VCR_RECORD_MODE", "once")
-api_vcr = vcr.VCR(
-    cassette_library_dir=f"{os.path.dirname(__file__)}/fixtures/isone/vcr_cassettes",
+api_vcr = setup_vcr(
+    source="isone",
     record_mode=record_mode,
-    match_on=["uri", "method"],
-    before_record=before_record_callback,
+    date_range_endpoints=DATE_RANGE_ENDPOINTS,
 )
